InterfaceLocoNet

+++ MyTrains/LocoSlotDataP2.swift
@@ -87,10 +87,10 @@ public class LocoSlotDataP2 : LocoNetMessage {
     }
   }
   
-  public var functions : Int {
+  public var functions : UInt64 {
     get {
       
-      var fnx : Int = 0
+      var fnx : UInt64 = 0
       
       var byte = message[9]
